issue #5: moved entities and DTOs in shared
