+++ .github/workflow/travlyn-ci.yml
@@ -1,0 +1,26 @@
+name: Travlyn CI
+
+on: [push]
+
+jobs:
+    build-server:
+        runs-on: ubuntu-latest
+
+        steps:
+        - uses: actions/checkout@v1
+        - name: Set up JDK 11
+          uses: actions/setup-java@v1
+          with:
+            java-version: 11
+        - name: Build with Maven
+          run: mvn clean install -DskipTests
+
+    build-client:
+        steps:
+        - uses: actions/checkout@v1
+        - name: set up JDK 1.8
+          uses: actions/setup-java@v1
+          with:
+            java-version: 1.8
+        - name: Build with Gradle
+          run: ./gradlew build
