+++ java/org.travlyn.server/pom.xml
@@ -30,6 +30,12 @@
     </build>
     <dependencies>
         <dependency>
+            <groupId>org.travlyn</groupId>
+            <artifactId>org.travlyn.shared</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
             <version>2.2.2.RELEASE</version>
@@ -39,17 +45,6 @@
             <artifactId>spring-boot-starter-tomcat</artifactId>
             <version>2.2.2.RELEASE</version>
         </dependency>
-        <!--SpringFox dependencies -->
-        <dependency>
-            <groupId>io.springfox</groupId>
-            <artifactId>springfox-swagger2</artifactId>
-            <version>2.9.2</version>
-        </dependency>
-        <dependency>
-            <groupId>io.springfox</groupId>
-            <artifactId>springfox-swagger-ui</artifactId>
-            <version>2.9.2</version>
-        </dependency>
 
         <dependency>
             <groupId>com.github.joschi.jackson</groupId>
@@ -57,11 +52,6 @@
             <version>2.6.4</version>
         </dependency>
         <!-- Bean Validation API support -->
-        <dependency>
-            <groupId>javax.validation</groupId>
-            <artifactId>validation-api</artifactId>
-            <version>2.0.1.Final</version>
-        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -94,13 +84,6 @@
             <version>8.0.17</version>
             <scope>compile</scope>
         </dependency>
-
-        <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.8.6</version>
-        </dependency>
-
         <dependency>
             <groupId>org.hsqldb</groupId>
             <artifactId>hsqldb</artifactId>
