+++ java/org.travlyn.server/src/main/java/org/travlyn/server/api/TripApiController.java
@@ -1,17 +1,14 @@
 package org.travlyn.server.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.swagger.annotations.ApiParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.travlyn.server.model.Rating;
-import org.travlyn.server.model.Trip;
+import org.travlyn.shared.model.api.Rating;
+import org.travlyn.shared.model.api.Trip;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
