+++ java/org.travlyn.server/src/main/java/org/travlyn/server/api/UserApiController.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.travlyn.server.model.Trip;
-import org.travlyn.server.model.User;
+import org.travlyn.shared.model.api.Trip;
+import org.travlyn.shared.model.api.User;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
