+++ java/org.travlyn.shared/src/main/java/org/travlyn/shared/model/api/AbstractDataTransferObject.java
@@ -1,4 +1,4 @@
-package org.travlyn.server.model;
+package org.travlyn.shared.model.api;
 
 public abstract class AbstractDataTransferObject implements DataTransferObject {
     @Override
