+++ java/org.travlyn.shared/src/main/java/org/travlyn/shared/model/api/Stop.java
@@ -1,10 +1,10 @@
-package org.travlyn.server.model;
+package org.travlyn.shared.model.api;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 import org.springframework.validation.annotation.Validated;
-import org.travlyn.server.db.model.RatingEntity;
-import org.travlyn.server.db.model.StopEntity;
+import org.travlyn.shared.model.db.RatingEntity;
+import org.travlyn.shared.model.db.StopEntity;
 
 import javax.validation.Valid;
 import java.util.*;
