+++ java/org.travlyn.shared/src/main/java/org/travlyn/shared/model/api/Token.java
@@ -1,9 +1,9 @@
-package org.travlyn.server.model;
+package org.travlyn.shared.model.api;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 import org.springframework.validation.annotation.Validated;
-import org.travlyn.server.db.model.TokenEntity;
+import org.travlyn.shared.model.db.TokenEntity;
 
 import java.util.Objects;
