+++ java/org.travlyn.shared/src/main/java/org/travlyn/shared/model/db/StopEntity.java
@@ -1,7 +1,6 @@
-package org.travlyn.server.db.model;
+package org.travlyn.shared.model.db;
 
 import javax.persistence.*;
-import java.util.List;
 import java.util.Set;
 
 @Entity
